pyproject.toml
[project]
name = "concord"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "redis>=5.0.7",
  "psycopg[binary]>=3.2.0",
  "sqlalchemy>=2.0.32",
]

[tool.uvicorn]
factory = false
